refactor : changement du type Double en BigDecimal pour eviter les erreurs de calcul

+++ src/main/java/com/mns/cda/saas_facturation/DTO/ArticleDTO.java
@@ -2,13 +2,16 @@ package com.mns.cda.saas_facturation.DTO;
 
 import com.mns.cda.saas_facturation.model.Tva;
 
+import java.math.BigDecimal;
+
+// Création d'un DTO qui pour tout ce qui va être affichage et sortie du Back pour le model Article
 public record ArticleDTO(
         Long artId,
         String artReference,
         String artName,
         String artDescription,
-        Double artPriceExcludeTaxes,
+        BigDecimal artPriceExcludeTaxes,
         int artStock,
         Tva tva,
-        Double artPriceTTC) {
+        BigDecimal artPriceTTC) {
 }
